Propose updating text for mixed content option
When this option is enabled using Mattermost on desktop and web has about the same risks. 

Therefore, propose we use the same defaults and also similar language? 

It would make the application potentially more usable by non-technical people. 

In future, this option might be pre-configured by an IT admin, and they can choose which default to use if it's an important part of their system security.