Fix webview doesn't get focus correctly
For #231

webview.focus() was used to focus on the webview, but it's DOM element function.
So only the tag was focused in the host window, and its webContents was not focused.
As the result, the shortcuts of Mattermost server didn't work in some cases.
Electron is upgraded to 1.2.8 because of webContents.isFocused().