Use build/ as CircleCI workspace
webpack compiles the application code to dist/ and electron-builder
treats it as a source of content to be packaged. As such, using this
path for CircleCI's workspace can cause unwanted scratch files to be
packaged into the app. By instead using build/, we solve the issue.
Moving the entitlements plist to the root of the repo is appropriate
because that's where all the other build config files currently live.

Fixes: #1809