Include more obvious prompt to reproduce issues in the Web App first
We've received a very large volume of issues to the Desktop App repository that are actually Web App issues. It would appear that users are missing/misinterpreting the prompt to try and reproduce in the Web App first.

This PR adds a much large prompt in the issue template to hopefully ensure reporter will know where the issues should go.